Rename back to thank-you.vue
