+++ myuw/static/vue/components/pages/index/cards/new_student/thank-you.vue
